Changed note from float to int.

+++ src/OPL3GM_proc.cpp
@@ -123,7 +123,7 @@ VstInt32 OPL3GM::processEvents (VstEvents* ev)
 				{
 					if (channel != 9)
 					{
-						float note = byte2;
+						int note = byte2;
 						note = note + (int)Transpose;
 						if (note > 127)
 						{
